+++ AIO_SER/src/main/java/com/mpri/aio/system/controller/SysDictController.java
@@ -43,7 +43,7 @@ public class SysDictController extends BaseController {
 	* @param sysDict
 	* @return
 	 */
-	@Logs(value = "编码查询")
+	@Logs(value = "编码查询",type ="QUERY")
 	@CrossOrigin
 	@PostMapping(value = "/list")
 	public PageIo<SysDict> list(int pageNo,int pageSize,SysDict sysDict) {
